+++ malcolm/modules/ca/parts/cachoicepart.py
@@ -1,6 +1,7 @@
 from malcolm.core import Part, PartRegistrar, ChoiceMeta, DEFAULT_TIMEOUT
 from ..util import CaToolsHelper, CAAttribute, APartName, AMetaDescription, \
-    APv, ARbv, ARbvSuff, AMinDelta, ATimeout, ADestinationPort, AWidget, AGroup, AConfig
+    APv, ARbv, ARbvSuffix, AMinDelta, ATimeout, ASinkPort, AWidget, \
+    AGroup, AConfig
 
 
 class CAChoicePart(Part):
@@ -11,10 +12,10 @@ class CAChoicePart(Part):
                  description,  # type: AMetaDescription
                  pv="",  # type: APv
                  rbv="",  # type: ARbv
-                 rbv_suff="",  # type: ARbvSuff
+                 rbv_suffix="",  # type: ARbvSuffix
                  min_delta=0.05,  # type: AMinDelta
                  timeout=DEFAULT_TIMEOUT,  # type: ATimeout
-                 inport=None,  # type: ADestinationPort
+                 sink_port=None,  # type: ASinkPort
                  widget=None,  # type: AWidget
                  group=None,  # type: AGroup
                  config=True,  # type: AConfig
@@ -24,8 +25,8 @@ class CAChoicePart(Part):
         catools = CaToolsHelper.instance()
         self.meta = ChoiceMeta(description)
         self.caa = CAAttribute(
-            self.meta, catools.DBR_ENUM, pv, rbv, rbv_suff, min_delta, timeout,
-            inport, widget, group, config, self.on_connect)
+            self.meta, catools.DBR_ENUM, pv, rbv, rbv_suffix, min_delta, timeout,
+            sink_port, widget, group, config, self.on_connect)
 
     def on_connect(self, value):
         self.meta.set_choices(value.enums)