+++ malcolm/modules/ca/parts/calongpart.py
@@ -1,6 +1,7 @@
 from malcolm.core import Part, PartRegistrar, NumberMeta, DEFAULT_TIMEOUT
 from ..util import CaToolsHelper, CAAttribute, APartName, AMetaDescription, \
-    APv, ARbv, ARbvSuff, AMinDelta, ATimeout, ADestinationPort, AWidget, AGroup, AConfig
+    APv, ARbv, ARbvSuffix, AMinDelta, ATimeout, ASinkPort, AWidget, \
+    AGroup, AConfig
 
 
 class CALongPart(Part):
@@ -11,10 +12,10 @@ class CALongPart(Part):
                  description,  # type: AMetaDescription
                  pv="",  # type: APv
                  rbv="",  # type: ARbv
-                 rbv_suff="",  # type: ARbvSuff
+                 rbv_suffix="",  # type: ARbvSuffix
                  min_delta=0.05,  # type: AMinDelta
                  timeout=DEFAULT_TIMEOUT,  # type: ATimeout
-                 inport=None,  # type: ADestinationPort
+                 sink_port=None,  # type: ASinkPort
                  widget=None,  # type: AWidget
                  group=None,  # type: AGroup
                  config=True,  # type: AConfig
@@ -24,7 +25,7 @@ class CALongPart(Part):
         catools = CaToolsHelper.instance()
         self.caa = CAAttribute(
             NumberMeta("int32", description), catools.DBR_LONG, pv, rbv,
-            rbv_suff, min_delta, timeout, inport, widget, group, config)
+            rbv_suffix, min_delta, timeout, sink_port, widget, group, config)
 
     def setup(self, registrar):
         # type: (PartRegistrar) -> None